refactor: ConnectionTemplate 이너클래스로 변경
